+++ dialogExecution/startVirtualMachine.py
@@ -419,7 +419,13 @@ class StartVirtualMachineDialog(QDialog, Ui_Dialog):
                 qemu_cmd = qemu_cmd + f" -cpu {self.vmSpecs[3]}"
 
             if self.vmSpecs[6] != "Let QEMU decide":
-                if self.vmSpecs[1] == "aarch64" or self.vmSpecs[1] == "arm":
+                if self.vmSpecs[6] == "std" or self.vmSpecs[6] == "qxl" or self.vmSpecs[6] == "cirrus" or self.vmSpecs[6] == "cg3" or self.vmSpecs[6] == "tcx":
+                    qemu_cmd = qemu_cmd + f" -vga {self.vmSpecs[6]}"
+
+                else:
+                    qemu_cmd = qemu_cmd + f" -device {self.vmSpecs[6]}"
+
+                """ if self.vmSpecs[1] == "aarch64" or self.vmSpecs[1] == "arm":
                     if self.vmSpecs[6] == "std":
                         qemu_cmd = qemu_cmd + f" -device VGA -display gtk"
                     
@@ -427,7 +433,7 @@ class StartVirtualMachineDialog(QDialog, Ui_Dialog):
                         qemu_cmd = qemu_cmd + f" -device {self.vmSpecs[6]} -display gtk"
 
                 else:
-                    qemu_cmd = qemu_cmd + f" -vga {self.vmSpecs[6]}"
+                    qemu_cmd = qemu_cmd + f" -vga {self.vmSpecs[6]}" """
 
             if self.vmSpecs[7] != "none":
                 if self.vmSpecs[1] == "i386" or self.vmSpecs[1] == "x86_64" or self.vmSpecs[1] == "ppc" or self.vmSpecs[1] == "ppc64" or self.vmSpecs[1] == "sparc" or self.vmSpecs[1] == "sparc64":