+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 
 - The New VM dialog is being optimised further.
 - An issue in the New VM dialog which caused the VM creation to be incomplete has been fixed.
+- The Start VM dialog has been adjusted.
 
 ## Known issues